Create weight and slot usability to be individual for each player
This new feature allows players to have different weights/slots from each other. This information is saved in the database along with the player's data.

With these new options, players can use the ChangeWeight and ChangeSlots methods to add or remove weight/slots from their characters.
It needs to be released along with the changes to rsg-core.

Signed-off-by: Eduardo Froehlich <eduardosilvafroehlich@gmail.com>